requirements.txt
streamlit>=1.37.0
requests>=2.31.0
pandas>=2.0.0
SQLAlchemy>=2.0.0
psycopg[binary]>=3.2.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.6.0
google-generativeai>=0.8.0
